+++ Assets/Tests/Runtime/ClientServer/ClientServerComponentTests.cs
@@ -83,7 +83,7 @@ namespace Mirage.Tests.Runtime.ClientServer
             var gameObject = new GameObject();
             NetworkIdentity identity = gameObject.AddComponent<NetworkIdentity>();
             identity.AssetId = guid;
-            identity.NetId = (uint)Random.Range(0, int.MaxValue);
+            identity.NetId = (ushort)Random.Range(0, ushort.MaxValue);
 
             clientObjectManager.RegisterSpawnHandler(guid, SpawnDelegateTest, go => { });
             clientObjectManager.RegisterPrefab(identity, guid);
@@ -102,7 +102,7 @@ namespace Mirage.Tests.Runtime.ClientServer
             var gameObject = new GameObject();
             NetworkIdentity identity = gameObject.AddComponent<NetworkIdentity>();
             identity.AssetId = guid;
-            identity.NetId = (uint)Random.Range(0, int.MaxValue);
+            identity.NetId = (ushort)Random.Range(0, ushort.MaxValue);
 
             UnSpawnDelegate unspawnDelegate = Substitute.For<UnSpawnDelegate>();
