+++ Assets/Tests/Runtime/ClientServer/GameObjectSyncvarTest.cs
@@ -62,7 +62,7 @@ namespace Mirage.Tests.Runtime.ClientServer
             serverObjectManager.Spawn(newObject);
 
             // wait until the client spawns it
-            uint newObjectId = newBehavior.NetId;
+            ushort newObjectId = newBehavior.NetId;
 
             NetworkIdentity newClientObject = await AsyncUtil.WaitUntilSpawn(client.World, newObjectId);
